style(components/items): move misc components to `/misc`
