+++ app/components/items/misc/item-image.skeleton.tsx
