+++ app/components/items/misc/item-image.spec.tsx
