+++ app/components/items/misc/item-image.stories.tsx
