+++ app/components/items/misc/item-image.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Image from 'next/image'
 import { ComponentProps } from 'react'
