+++ app/components/items/misc/item-position.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { HTMLAttributes } from 'react'
 
 import { cn } from '@app/utils/cn'
